+++ lib/utils/years_old.dart
@@ -1,0 +1,12 @@
+int? yearsOld(DateTime? birthDate) {
+  // Parsed date to check
+
+  if (birthDate == null) {
+    return null;
+  }
+
+  // Date to check but moved 18 years ahead
+  DateTime adultDate = DateTime.now();
+
+  return ((adultDate.difference(birthDate).inDays) / 365).floor();
+}
