+++ lib/widgets/cert_spimplified_info.dart
@@ -1,6 +1,7 @@
 import 'package:circle_flags/circle_flags.dart';
 import 'package:covid_checker/generated/l10n.dart';
 import 'package:covid_checker/models/result.dart';
+import 'package:covid_checker/utils/years_old.dart';
 import 'package:covid_checker/widgets/cert_detailed_view.dart';
 import 'package:covid_checker/widgets/detail.dart';
 import 'package:dart_cose/dart_cose.dart';
@@ -119,17 +120,4 @@ class CertInfoViewer extends StatelessWidget {
       ),
     );
   }
-}
-
-int? yearsOld(DateTime? birthDate) {
-  // Parsed date to check
-
-  if (birthDate == null) {
-    return null;
-  }
-
-  // Date to check but moved 18 years ahead
-  DateTime adultDate = DateTime.now();
-
-  return ((adultDate.difference(birthDate).inDays) / 365).floor();
 }
